Extend datasource to access log data from monasca-log-api.
Extend the datasource to be able to return what Grafana refers to as
'doc' objects (essentially JSON) from the monasca-log-api. This will
allow viewing of log entries in Table panels side-by-side with metrics
in Graph or other panels. Due to the way Grafana handles datasource
endpoints, one datasource cannot access multiple endpoints, so two
instances of the plugin must be created. One for metrics, and one for
logs. This change should not break existing installations, and will
assume it's previous behaviour (metrics) unless explicitly configured.